Add the payment details for manager

+++ routes/web.php
@@ -23,6 +23,7 @@ use App\Http\Controllers\NewMemberController;
 use App\Http\Controllers\DeleteRequestController;
 use App\Http\Controllers\EventUpdateRequestController;
 use App\Http\Controllers\EventHostPaymentController;
+use App\Http\Controllers\PaymentReportController;
 
 
 
@@ -70,9 +71,9 @@ Route::get('/reviews', function () {
 //     return Inertia::render('Cart/TBCart/TBCart');
 // })->name('buyticketscart');
 
-Route::get('/dashboard', function () {
-    return Inertia::render('Dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
+// Route::get('/dashboard', function () {
+//     return Inertia::render('Dashboard');
+// })->middleware(['auth', 'verified'])->name('dashboard');
 
 // Manager
 Route::get('/manager', function () {
@@ -135,7 +136,7 @@ Route::get('programmer/dashboard', [ProgrammerController::class, 'index'])->name
 
 Route::get('admin/dashboard', [AdminController::class, 'index'])->name('admin.dashboard')->middleware(['auth', 'admin']);
 
-Route::get('eventhost/dashboard', [EventHostController::class, 'index'])->name('eventhost.dashboard')->middleware(['auth', 'eventhost']);
+Route::get('eventhost/dashboard', [EventHostController::class, 'index'])->name('eventhost.dashboard');
 
 
 Route::get('eventhost/ongoing', [EventHostController::class, 'ehOngoing'])->name('eventhost.ongoingEvents')->middleware(['auth', 'eventhost']);
@@ -148,14 +149,14 @@ Route::get('eventhost/profile', [EventHostController::class, 'profile'])->name('
 
 
 
-Route::get('event/create', [EventController::class, 'create'])->name('event.create');
+Route::get('event/create', [EventController::class, 'create'])->name('event.create')->middleware(['auth', 'eventhost']);
 Route::post('event/store', [EventController::class, 'store'])->name('event.store');
 
 Route::get('/api/events', [EventController::class, 'getCompletedEvents']);
 
 Route::get('/api/get-event/{event}', [EventController::class, 'getEvent']);
 
-Route::middleware(['auth'])->group(function () {
+Route::middleware(['auth','eventhost'])->group(function () {
     Route::get('/event-host/profile', [EventHostController::class, 'profile'])->name('eventhost.profile');
     Route::get('/event-host/history', [EventHostController::class, 'ehOngoing'])->name('eventhost.history');
     Route::get('/event-host/changePW', [EventHostController::class, 'changePW'])->name('eventhost.changePW');
@@ -164,11 +165,11 @@ Route::middleware(['auth'])->group(function () {
 });
 
 
-Route::get('/EHOngoing', [EventHostController::class, 'ehOngoing'])->name('eh.ongoing');
-Route::get('/EHPendingPayments', [EventHostController::class, 'ehPendingPayments'])->name('eh.pendingPayments');
-Route::get('/EHPendingRequests', [EventHostController::class, 'ehPendingRequests'])->name('eh.pendingRequests');
-Route::get('/EHRejected', [EventHostController::class, 'ehRejected'])->name('eh.rejected');
-Route::get('/EHHistory', [EventHostController::class, 'ehHistory'])->name('eh.history');
+Route::get('/EHOngoing', [EventHostController::class, 'ehOngoing'])->name('eh.ongoing')->middleware(['auth', 'eventhost']);
+Route::get('/EHPendingPayments', [EventHostController::class, 'ehPendingPayments'])->name('eh.pendingPayments')->middleware(['auth', 'eventhost']);
+Route::get('/EHPendingRequests', [EventHostController::class, 'ehPendingRequests'])->name('eh.pendingRequests')->middleware(['auth', 'eventhost']);
+Route::get('/EHRejected', [EventHostController::class, 'ehRejected'])->name('eh.rejected')->middleware(['auth', 'eventhost']);
+Route::get('/EHHistory', [EventHostController::class, 'ehHistory'])->name('eh.history')->middleware(['auth', 'eventhost']);
 
 // Get attendee count
 Route::get('/api/event/{id}/attendees', [EventController::class, 'getAttendees']);
@@ -188,9 +189,9 @@ Route::post('/artists', [ArtistController::class, 'store']);
 Route::get('/eventdetails/{id}', [EventController::class, 'show'])->name('event.details');
 Route::get('/buytickets/{id}', [EventController::class, 'showBuyTickets'])->name('buytickets');
 
-Route::get('/event/create-requests', [EventController::class, 'getPendingEvents'])->name('event.create.requests');
+Route::get('/event/create-requests', [EventController::class, 'getPendingEvents'])->name('event.create.requests')->middleware(['auth', 'adminManagerProgrammer']);
 
-Route::post('/manager/approve-event/{id}', [EventController::class, 'approveEvent'])->name('manager.approve.event');
+Route::post('/manager/approve-event/{id}', [EventController::class, 'approveEvent'])->name('manager.approve.event')->middleware(['auth', 'adminManagerProgrammer']);
 
 Route::get('/manager/delete-event/{id}', [EventController::class, 'deleteEvent'])->name('manager.delete.event');
 
@@ -240,11 +241,11 @@ Route::post('/payment/cancel/update', [StripePaymentController::class, 'updateCa
 Route::get('/payment/cancel', [StripePaymentController::class, 'showCancelPage'])
     ->name('payment.cancel');
 
-Route::get('ongoing',[ManagerController::class,'showManagerOngoingEvents'])->name('ongoing');
+Route::get('ongoing',[ManagerController::class,'showManagerOngoingEvents'])->name('ongoing')->middleware(['auth', 'adminManagerProgrammer']);
 
 Route::get('view-event/{event}',[ManagerController::class,'showManagerViewEvent'])->name('viewevent');
 
-Route::get('ended-event-history', [EventController::class, 'getEndedEvents'])->name('ended.event.history');
+Route::get('ended-event-history', [EventController::class, 'getEndedEvents'])->name('ended.event.history')->middleware(['auth', 'adminManagerProgrammer']);
 
 Route::get('pending-payment-history', [EventController::class, 'getPendingPaymentEvents'])->name('pending.payments.history');
 
@@ -257,7 +258,7 @@ Route::get('/ended-event/{id}/report', [EventController::class, 'generateEndedEv
 
 Route::get('/rejected-event/{id}/report', [EventController::class, 'generateRejectedEventReport'])->name('rejected.event.report');
 
-Route::middleware(['auth'])->group(function () {
+Route::middleware(['auth','adminManagerProgrammer'])->group(function () {
 Route::get('/other/profile-show', [ManagerController::class, 'showManagerProfile'])->name('manager.profile');
 });
 
@@ -281,20 +282,20 @@ Route::post('/reviews', [WebSiteReviewController::class, 'store'])->name('review
 
 Route::get('/reviewPg', [WebSiteReviewController::class, 'showReviewPage'])->name('reviews.show');
 
-Route::get('/show-reviews', [WebSiteReviewController::class, 'index'])->middleware('auth')->name('reviews.index');
+Route::get('/show-reviews', [WebSiteReviewController::class, 'index'])->middleware('auth')->name('reviews.index')->middleware(['auth', 'adminManagerProgrammer']);
 
-Route::post('/reviews/{id}/approve', [WebSiteReviewController::class, 'approve'])->middleware('auth');
+Route::post('/reviews/{id}/approve', [WebSiteReviewController::class, 'approve'])->middleware('auth')->middleware(['auth', 'adminManagerProgrammer']);
 
-Route::post('/reviews/{id}/reply', [WebSiteReviewController::class, 'reply'])->middleware('auth');
+Route::post('/reviews/{id}/reply', [WebSiteReviewController::class, 'reply'])->middleware('auth')->middleware(['auth', 'adminManagerProgrammer']);
 
-Route::delete('/reviews/{id}', [WebSiteReviewController::class, 'destroy'])->middleware('auth');
-
-
+Route::delete('/reviews/{id}', [WebSiteReviewController::class, 'destroy'])->middleware('auth')->middleware(['auth', 'adminManagerProgrammer']);
 
 
-    Route::get('/inquiries', [InquiryController::class, 'index'])->name('inquiries.index');
-    Route::post('/inquiries/{id}/reply', [InquiryController::class, 'reply'])->name('inquiries.reply');
-    Route::delete('/inquiries/{id}', [InquiryController::class, 'destroy'])->name('inquiries.destroy');
+
+
+    Route::get('/inquiries', [InquiryController::class, 'index'])->name('inquiries.index')->middleware(['auth', 'adminManagerProgrammer']);
+    Route::post('/inquiries/{id}/reply', [InquiryController::class, 'reply'])->name('inquiries.reply')->middleware(['auth', 'adminManagerProgrammer']);
+    Route::delete('/inquiries/{id}', [InquiryController::class, 'destroy'])->name('inquiries.destroy')->middleware(['auth', 'adminManagerProgrammer']);
 
 // Public inquiry submission route
 Route::post('/inquiries', [InquiryController::class, 'store'])->name('inquiries.store');
@@ -303,15 +304,15 @@ Route::get('/connect-with-us', function () {
     return Inertia::render('ConnectWithUs/ConnectWithUs');
 })->name('connect-with-us');
 
-Route::get('/manager/stats', [ManagerController::class, 'getStats']);
+Route::get('/manager/stats', [ManagerController::class, 'getStats'])->middleware(['auth', 'adminManagerProgrammer']);
 
-Route::get('/add-new-member', [NewMemberController::class, 'index'])->name('add.new.memeber');
+Route::get('/add-new-member', [NewMemberController::class, 'index'])->name('add.new.memeber')->middleware(['auth', 'adminManagerProgrammer']);
 
 Route::post('/register-new-member', [NewMemberController::class, 'store']);
 
 Route::post('/event/delete-request', [EventController::class, 'storeDeleteRequest']);
 
-Route::middleware(['auth'])->group(function () {
+Route::middleware(['auth','customer'])->group(function () {
     Route::get('/tb-profile', function () {
         return Inertia::render('UserProfile/TB-Profiles/TBProfile', [ 
             'user' => auth()->user()
@@ -343,18 +344,18 @@ Route::get('/change-password', [UserController::class, 'showChangePasswordForm']
 Route::post('/change-password', [UserController::class, 'updatePassword'])->name('user.change-password');
 
 
-Route::get('other/profile', [AdminController::class, 'profile'])->name('admin.profile');
-Route::post('other/updateProfile', [AdminController::class, 'updateProfile'])->name('admin.updateProfile');
+Route::get('other/profile', [AdminController::class, 'profile'])->name('admin.profile')->middleware(['auth', 'adminManagerProgrammer']);
+Route::post('other/updateProfile', [AdminController::class, 'updateProfile'])->name('admin.updateProfile')->middleware(['auth', 'adminManagerProgrammer']);
 Route::get('other/changePW', [AdminController::class, 'changePW'])->name('admin.changePW');
-Route::post('other/update-password', [AdminController::class, 'updatePassword'])->name('admin.updatePassword');
-Route::get('other/signOut', [AdminController::class, 'signOut'])->name('admin.signOut');
+Route::post('other/update-password', [AdminController::class, 'updatePassword'])->name('admin.updatePassword')->middleware(['auth', 'adminManagerProgrammer']);
+Route::get('other/signOut', [AdminController::class, 'signOut'])->name('admin.signOut')->middleware(['auth', 'adminManagerProgrammer']);
 
-Route::get('other/ongoing', [AdminController::class, 'ongoingEvents'])->name('admin.ongoingEvents');
-Route::get('other/userDetails', [AdminController::class, 'userDetails'])->name('admin.userDetails');
+Route::get('other/ongoing', [AdminController::class, 'ongoingEvents'])->name('admin.ongoingEvents')->middleware(['auth', 'adminManagerProgrammer']);
+Route::get('other/userDetails', [AdminController::class, 'userDetails'])->name('admin.userDetails')->middleware(['auth', 'adminManagerProgrammer']);
 
 Route::get('/events/view/{id}', [AdminController::class, 'show'])->name('admin.viewEvent');
 
-Route::get('/addMember', [ProgrammerController::class, 'addMember'])->name('add.member');
+Route::get('/addMember', [ProgrammerController::class, 'addMember'])->name('add.member')->middleware(['auth', 'adminManagerProgrammer']);
 
 Route::get('/events/search', [EventController::class, 'search'])->name('events.search');
  //Route::get('/events/details', [EventController::class, 'searchDetails'])->name('events.details');
@@ -370,58 +371,69 @@ Route::post('/event/exists', [EventController::class, 'eventExists']);
 
  Route::get('/events/{id}/delete', [EventController::class, 'delete'])->name('delete.event');
 
- Route::get('other/tbUserDetails', [AdminController::class, 'tbUserDetails'])->name('tb.userDetails');
+ Route::get('other/tbUserDetails', [AdminController::class, 'tbUserDetails'])->name('tb.userDetails')->middleware(['auth', 'adminProgrammer']);
 
- Route::get('other/ehUserDetails', [AdminController::class, 'ehUserDetails'])->name('eh.userDetails');
+ Route::get('other/ehUserDetails', [AdminController::class, 'ehUserDetails'])->name('eh.userDetails')->middleware(['auth', 'adminProgrammer']);
 
- Route::post('delete/users/{id}', [AdminController::class, 'destroy'])->name('users.destroy');
+ Route::post('delete/users/{id}', [AdminController::class, 'destroy'])->name('users.destroy')->middleware(['auth', 'adminProgrammer']);
 
- Route::get('/users/{id}/download', [AdminController::class, 'ticketBuyerHistory'])->name('users.download');
+ Route::get('/users/{id}/download', [AdminController::class, 'ticketBuyerHistory'])->name('users.download')->middleware(['auth', 'adminProgrammer']);
 
- Route::get('/users/{id}/download1', [AdminController::class, 'eventHostHistory'])->name('users.download1');
+ Route::get('/users/{id}/download1', [AdminController::class, 'eventHostHistory'])->name('users.download1')->middleware(['auth', 'adminProgrammer']);
 
 Route::get('/event/update/{event_id}', [EventUpdateRequestController::class, 'edit'])->name('event.update');
 
 Route::post('/event/update/{event_id}', [EventUpdateRequestController::class, 'store'])->name('event.update.store');
 
-Route::get('/event/update-requests', [EventUpdateRequestController::class, 'index'])->name('event.update.requests');
+Route::get('/event/update-requests', [EventUpdateRequestController::class, 'index'])->name('event.update.requests')->middleware(['auth', 'adminManagerProgrammer']);
 
-Route::post('/event/update/approve/{event_update_request_id}', [EventUpdateRequestController::class, 'approve'])->name('event.update.approve');
+Route::post('/event/update/approve/{event_update_request_id}', [EventUpdateRequestController::class, 'approve'])->name('event.update.approve')->middleware(['auth', 'adminManagerProgrammer']);
 
-Route::get('/event/selling-ticket-report/{id}', [EventHostController::class, 'generateTicketReport']);
+Route::get('/event/selling-ticket-report/{id}', [EventHostController::class, 'generateTicketReport'])->middleware(['auth', 'eventhost']);
 
-Route::get('/event/event-attendees-list/{id}', [EventHostController::class, 'generateAttendeesList']);
+Route::get('/event/event-attendees-list/{id}', [EventHostController::class, 'generateAttendeesList'])->middleware(['auth', 'eventhost']);
 
-Route::post('/event/cancel/{id}', [EventHostController::class, 'cancelEvent']);
+Route::post('/event/cancel/{id}', [EventHostController::class, 'cancelEvent'])->middleware(['auth', 'eventhost']);
 
-Route::get('/event/rejection-pdf/{id}', [EventHostController::class, 'generateRejectionPDF'])->name('event.rejection-pdf');
+Route::get('/event/rejection-pdf/{id}', [EventHostController::class, 'generateRejectionPDF'])->name('event.rejection-pdf')->middleware(['auth', 'eventhost']);
 
 
 Route::get('/ended-event/report/{id}', [EventHostController::class, 'downloadEventReport'])
-    ->name('event.report.download');
+    ->name('event.report.download')->middleware(['auth', 'eventhost']);
 
     Route::prefix('eventhost/payment')->group(function () {
         Route::get('/checkout', [EventHostPaymentController::class, 'checkout'])->name('eventhost.payment.checkout');
         Route::post('/process', [EventHostPaymentController::class, 'processPayment'])->name('eventhost.payment.process');
         Route::get('/success', [EventHostPaymentController::class, 'success'])->name('eventhost.payment.success');
         Route::get('/cancel', [EventHostPaymentController::class, 'cancel'])->name('eventhost.payment.cancel');
-    });
+    })->middleware(['auth', 'eventhost']);
 
 
-    Route::get('eventhost/profile/edit', [EventHostController::class, 'editProfile'])->name('eventhost.profile.edit');
+    Route::get('eventhost/profile/edit', [EventHostController::class, 'editProfile'])->name('eventhost.profile.edit')->middleware(['auth', 'eventhost']);
 
-    Route::post('/eventhost/profile/update', [EventHostController::class, 'updateProfile'])->name('eventhost.profile.update');
+    Route::post('/eventhost/profile/update', [EventHostController::class, 'updateProfile'])->name('eventhost.profile.update')->middleware(['auth', 'eventhost']);
 
-    Route::post('/eventhost/profile/update-password', [EventHostController::class, 'updateEventHostPassword'])->name('eventhost.profile.update-password');
+    Route::post('/eventhost/profile/update-password', [EventHostController::class, 'updateEventHostPassword'])->name('eventhost.profile.update-password')->middleware(['auth', 'eventhost']);
 
 
 
-Route::get('/admin/monthly-user-registrations-by-role', [AdminController::class, 'getMonthlyUserRegistrationsByRole']);
+Route::get('/admin/monthly-user-registrations-by-role', [AdminController::class, 'getMonthlyUserRegistrationsByRole'])->middleware(['auth', 'admin']);
 
     Route::get('/host-contact/{id}', [EventController::class, 'showPage'])->name('host.contact.page');
     Route::post('/host-contact', [EventController::class, 'send'])->name('host.contact');
 
-Route::get('/admin/monthly-user-registrations-by-role', [AdminController::class, 'getMonthlyUserRegistrationsByRole']);
+Route::get('/admin/monthly-user-registrations-by-role', [AdminController::class, 'getMonthlyUserRegistrationsByRole'])->middleware(['auth', 'adminProgrammer']);
+
+Route::get('/paymenthistory', [PaymentReportController::class, 'index'])->name('payment.details')->middleware(['auth', 'adminManagerProgrammer']);
+
+
+
+Route::get('/payment-details/pdf', [PaymentReportController::class, 'exportPdf'])->name('payment.details.pdf')->middleware(['auth', 'adminManagerProgrammer']);
+
+Route::get('/customer/edit-profile', [UserController::class, 'editProfile'])->name('customer.editProfile')->middleware(['auth', 'customer']);
+
+Route::post('/customer/updateCustomerProfile',[UserController::class,'updateProfile'])->name('customer.updateProfile')->middleware(['auth', 'customer']);
+
 
 
 require __DIR__ . '/auth.php';
